unit-tests/mqtt_uplink_test/requirements.txt
amqp==2.3.2
aniso8601==3.0.0
anyjson==0.3.3
APScheduler==3.5.1
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
circus==0.15.0
circus-web==1.0.0
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
hiredis==0.2.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
paho-mqtt==1.3.1
pika==0.11.2
protobuf==3.5.1
psutil==5.4.6
pymongo==3.6.1
pytz==2018.4
pyzmq==16.0.4
redis==2.10.6
requests==2.19.0
simplejson==3.15.0
six==1.11.0
tomako==0.1.0.post1
TornadIO2==0.0.3
tornado==4.5.3
tzlocal==1.5.1
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
